requirements.txt
gradio
python-dotenv
langchain
langchain-groq
chromadb
sentence-transformers
PyPDF2
torch

